cleaned up the feature files
