+++ TestConfigAnimalSelectPage.properties
@@ -1,0 +1,4 @@
+browser = firefox
+baseUrl = http://www.thetestroom.com/jswebapp/animalselection.html
+mastertestdatafile = MasterTestDataFile.xls
+testdatasheet = TestCaseDataSets
