+++ src/test/resource/SubmitContact.feature
@@ -2,7 +2,7 @@ Feature: Submit a valid contact form use dataTable
 
 	@Application
   Scenario: Submit form using valid data
-    Given I am on the zoo site
+    Given I am on zoo site
     When I navigate to contact
     And I submit the form with valid data
       | Fields   | Value                  |
