+++ src/test/resource/adoptionScenarioHook.feature
@@ -1,6 +1,6 @@
 Feature: Adopt an animal Scenario Hooks
 
-  @web
+  @TestRoom
   Scenario: Adopt an available animal
     Given I am at the zoo site
     When I navigate to adoption
