+++ src/test/resource/selectAnimalFromAnimalSelectPage.feature
@@ -1,5 +1,6 @@
 Feature: To test Simba Lion can be selected in the Animal page.
 
+	@AnimalSelect
   Scenario: Navigate to Animal Selection page and select Simba Lion.
     Given I am on Animal Select page
     When I select Simba the Lion
